+++ lib/db_fuel/library.rb
@@ -10,6 +10,8 @@
 require_relative 'library/active_record/find_or_insert'
 require_relative 'library/active_record/insert'
 require_relative 'library/active_record/update'
+require_relative 'library/active_record/update_all'
+require_relative 'library/active_record/upsert'
 
 require_relative 'library/dbee/query'
 require_relative 'library/dbee/range'
@@ -20,6 +22,8 @@ module Burner
     register 'db_fuel/active_record/find_or_insert', DbFuel::Library::ActiveRecord::FindOrInsert
     register 'db_fuel/active_record/insert',         DbFuel::Library::ActiveRecord::Insert
     register 'db_fuel/active_record/update',         DbFuel::Library::ActiveRecord::Update
+    register 'db_fuel/active_record/update_all',     DbFuel::Library::ActiveRecord::UpdateAll
+    register 'db_fuel/active_record/upsert',         DbFuel::Library::ActiveRecord::Upsert
 
     register 'db_fuel/dbee/query',                   DbFuel::Library::Dbee::Query
     register 'db_fuel/dbee/range',                   DbFuel::Library::Dbee::Range