+++ lib/db_fuel/library/active_record/update.rb
@@ -7,22 +7,22 @@
 # LICENSE file in the root directory of this source tree.
 #
 
-require_relative 'base'
+require_relative 'upsert'
 
 module DbFuel
   module Library
     module ActiveRecord
-      # This job can take the objects in a register and updates them within database table.
+      # This job can take the unique objects in a register and updates them within database table.
       # The attributes translate to SQL SET clauses and the unique_keys translate to
-      # WHERE clauses.
+      # WHERE clauses to find the records to update.
+      # The primary_keyed_column is used to update the unique record.
+      # Only one record will be updated per statement.
       #
       # Expected Payload[register] input: array of objects
       # Payload[register] output: array of objects.
-      class Update < Base
-        attr_reader :unique_attribute_renderers
-
+      class Update < Upsert
         # Arguments:
-        #   name [required]: name of the job within the Burner::Pipeline.
+        #   name: name of the job within the Burner::Pipeline.
         #
         #   table_name [required]: name of the table to use for the INSERT statements.
         #
@@ -35,6 +35,11 @@ module DbFuel
         #          returned objects will be printed in the output.  Only use this option while
         #          debugging issues as it will fill up the output with (potentially too much) data.
         #
+        #   primary_keyed_column [required]: Primary key column for the corresponding table.
+        #                           Used as the WHERE clause for the UPDATE statement.
+        #                           Only one record will be updated at a time
+        #                           using the primary key specified.
+        #
         #   separator: Just like other jobs with a 'separator' option, if the objects require
         #              key-path notation or nested object support, you can set the separator
         #              to something non-blank (like a period for notation in the
@@ -43,32 +48,34 @@ module DbFuel
         #   timestamps: If timestamps is true (default behavior) then the updated_at column will
         #               automatically have its value set to the current UTC timestamp.
         #
-        #   unique_attributes: Each key will become a WHERE clause in order to only update specific
-        #                      records.
+        #   unique_attributes: Each key will become a WHERE clause in order to only find specific
+        #                      records. The UPDATE statement's WHERE
+        #                      clause will use the primary key specified.
         def initialize(
-          name:,
           table_name:,
+          name: '',
           attributes: [],
           debug: false,
+          primary_keyed_column: nil,
           register: Burner::DEFAULT_REGISTER,
           separator: '',
           timestamps: true,
           unique_attributes: []
         )
-          explicit_attributes = Burner::Modeling::Attribute.array(attributes)
 
-          attributes = timestamps ? timestamp_attributes + explicit_attributes : explicit_attributes
+          attributes = Burner::Modeling::Attribute.array(attributes)
 
           super(
             name: name,
             table_name: table_name,
             attributes: attributes,
             debug: debug,
+            primary_keyed_column: primary_keyed_column,
             register: register,
-            separator: separator
+            separator: separator,
+            timestamps: timestamps,
+            unique_attributes: unique_attributes
           )
-
-          @unique_attribute_renderers = make_attribute_renderers(unique_attributes)
 
           freeze
         end
@@ -79,14 +86,11 @@ module DbFuel
           payload[register] = array(payload[register])
 
           payload[register].each do |row|
-            set_object   = transform(attribute_renderers, row, payload.time)
-            where_object = transform(unique_attribute_renderers, row, payload.time)
+            rows_affected = 0
 
-            sql = db_provider.update_sql(set_object, where_object)
+            first_record = update_record(output, row, payload.time)
 
-            debug_detail(output, "Update Statement: #{sql}")
-
-            rows_affected = db_provider.update(set_object, where_object)
+            rows_affected = 1 if first_record
 
             debug_detail(output, "Individual Rows Affected: #{rows_affected}")
 
@@ -94,14 +98,6 @@ module DbFuel
           end
 
           output.detail("Total Rows Affected: #{total_rows_affected}")
-        end
-
-        private
-
-        def timestamp_attributes
-          [
-            updated_at_timestamp_attribute
-          ]
         end
       end
     end