+++ spec/db_fuel/library/active_record/find_or_insert_spec.rb
@@ -29,7 +29,7 @@ describe DbFuel::Library::ActiveRecord::FindOrInsert do
         { key: :last_name }
       ],
       table_name: 'patients',
-      primary_key: { key: :id },
+      primary_keyed_column: { key: :id },
       unique_attributes: [
         { key: :chart_number }
       ]
@@ -91,7 +91,7 @@ describe DbFuel::Library::ActiveRecord::FindOrInsert do
       expect(output.outs.first.string).to include('Total Inserted: 2')
     end
 
-    it 'sets primary_key for all payload objects' do
+    it 'sets primary_keyed_column for all payload objects' do
       payload[register].each do |object|
         expected = Patient.find_by(chart_number: object['chart_number']).id
         actual   = object['id']