+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 2.0.0 (March 10th, 2021)
+
+New Jobs:
+* db_fuel/active_record/upsert
+* db_fuel/active_record/update_all
+
+Changes:
+* db_fuel/active_record/update now only updates a single record. Use db_fuel/active_record/update_all to update multiple records at a time.
+
 # 1.1.0 (Decmeber 1st, 2020)
 
 New Jobs: